Compile `ldd as a module` in Buildroot:
Without `LDD_MODULE_MAKE_OPTS = KERNELDIR=$(LINUX_DIR)`, the module would not be able to compile because it won't know where the kernel headers are.

Without `$(eval $(kernel-module))`, Buildroot won't treat the package as a kernel module, meaning the module won’t be built and installed properly in the final image.